+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Don't include deleted config in `keys()` etc.
+
 ## 0.1.17 - 2024-01-18
 
 - Don't report empty contexts with empty keys. Throw if required environment variable is absent. Throw if we can't parse an environment variable correctly. (#63)